cloud_common/src/main/java/com/lew/common/util/JacksonUtil.java
/*
 * Copyright (c) 2023. This is my custom copyright information.2023-2023. All rights reserved.
 */

package com.lew.common.util;

import com.fasterxml.jackson.annotation.JsonInclude.Include;
import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.core.type.TypeReference;
import com.fasterxml.jackson.databind.DeserializationFeature;
import com.fasterxml.jackson.databind.JavaType;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.SerializationFeature;
import lombok.extern.slf4j.Slf4j;
import org.apache.commons.lang3.StringUtils;

import java.io.IOException;
import java.util.Objects;

/**
 * @author Yolen
 * @date 2023/7/25
 */
@Slf4j
public class JacksonUtil {
    public static final ObjectMapper MAPPER = new ObjectMapper();

    private JacksonUtil() {
    }

    static {
        // 设置输入:禁止把POJO中值为null的字段映射到json字符串中(注意：只对VO起作用；对Map List不起作用)
        MAPPER.setSerializationInclusion(Include.NON_NULL)
            // 忽略未知字段: 反序列化时，属性不存在的兼容处理
            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
            // 忽略空Bean转json的错误
            .configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false);
    }

    /**
     * 对象转Json格式字符串 参考：<a href="https://www.cnblogs.com/christopherchan/p/11071098.html">利用Jackson封装常用JsonUtil工具类</a>
     *
     * @param obj 对象
     * @return Json格式字符串
     */
    public static <T> String obj2String(T obj) {
        if (Objects.isNull(obj)) {
            return null;
        }
        try {
            return obj instanceof String ? (String)obj : MAPPER.writeValueAsString(obj);
        } catch (JsonProcessingException e) {
            log.warn("failed to parse Object to String : {}", e.getMessage());
            return null;
        }
    }

    /**
     * 字符串转换为自定义对象
     *
     * @param str 要转换的字符串
     * @param clazz 自定义对象的class对象
     * @return 自定义对象
     */
    public static <T> T string2Obj(String str, Class<T> clazz) {
        if (StringUtils.isBlank(str) || Objects.isNull(clazz)) {
            return null;
        }
        try {
            return clazz.equals(String.class) ? (T)str : MAPPER.readValue(str, clazz);
        } catch (Exception e) {
            log.warn("Parse String to Object error : {}", e.getMessage());
            return null;
        }
    }

    public static <T> T string2Obj(String str, TypeReference<T> typeReference) {
        if (StringUtils.isEmpty(str) || typeReference == null) {
            return null;
        }
        try {
            return (T)(typeReference.getType().equals(String.class) ? str : MAPPER.readValue(str, typeReference));
        } catch (IOException e) {
            log.warn("Parse String to Object error", e);
            return null;
        }
    }

    public static <T> T str2Collection(String str, Class<?> collectionClazz, Class<?>... elementClazzes) {
        JavaType javaType = MAPPER.getTypeFactory().constructParametricType(collectionClazz, elementClazzes);
        try {
            return MAPPER.readValue(str, javaType);
        } catch (IOException e) {
            log.warn("Parse String to Object error : {}" + e.getMessage());
            return null;
        }
    }
}
